requirements.txt
python-dotenv

#latest as of 03-08-2025
jupyter==1.1.1
notebook==7.3.2
ipykernel==6.29.5

langgraph
langsmith
langchain_anthropic
langchain_openai
langchain_community

tavily-python 
